0.2.0 - Created BashpileAst, antlr generates visitors
